review fix and add dry run

+++ scripts/delete_resources.sh
@@ -2,15 +2,31 @@
 
 delete_ec2_instances () {
   EC2_INSTANCE_IDS=$(aws ec2 describe-instances \
-    --filters "Name=tag:Name,Values=$1*" \
-    "Name=instance-state-name,Values=running" --query \
-    'Reservations[].Instances[].InstanceId' --output text)
-  if [ "${EC2_INSTANCE_IDS}" = "" ];then
-    echo "No ec2 instances found with prefix: $1. Nothing to delete."
+    --filters "Name=tag:Name,Values=$1*" "Name=instance-state-name,Values=running" \
+    --query 'Reservations[].Instances[].InstanceId' --output text)
+  if [ -z "${DRY_RUN}" ]; then
+    if [ "${EC2_INSTANCE_IDS}" = "" ];then
+      echo "No ec2 instances found with prefix: $1. Nothing to delete."
+    else
+      echo "Terminating ec2 instances for $1 if still up and running:
+      INSTANCE IDs: ${EC2_INSTANCE_IDS}"
+      for INSTANCE_ID in ${EC2_INSTANCE_IDS}
+      do
+        echo "Deleting instance id: ${INSTANCE_ID}"
+        aws ec2 terminate-instances --instance-ids "${INSTANCE_ID}" > /dev/null 2>&1
+      done
+    fi
   else
-    echo "Terminating ec2 instances for $1 if still up and running:"
-    # shellcheck disable=SC2046
-    aws ec2 terminate-instances --instance-ids "${EC2_INSTANCE_IDS}" > /dev/null 2>&1
+    echo "EC2 instances matching tag name for the prefix provided $1:"
+    EC2_TAG_NAMES=$(aws ec2 describe-instances \
+    --filters "Name=tag:Name,Values=archk3s*" "Name=instance-state-name,Values=running" \
+    --query 'Reservations[].Instances[].Tags' --output text)
+    echo "${EC2_TAG_NAMES}"
+    echo "Instance ID List: ${EC2_INSTANCE_IDS}"
+    TAG_COUNT=$(echo $EC2_TAG_NAMES |xargs -n1 echo | wc -l)
+    echo "EC2 Tag Name count: $((TAG_COUNT/2))"
+    INSTANCE_COUNT=$(echo $EC2_INSTANCE_IDS | xargs -n1 echo | wc -l)
+    echo "EC2 Instance Id count:$INSTANCE_COUNT"
   fi
 }
 
@@ -22,10 +38,16 @@ delete_db_resources () {
   if [ "${DB_INSTANCES}" = "" ];then
     echo "No db instances found with prefix $1. Nothing to delete."
   else
-    echo "Deleting db instances for $1:"
-    for INSTANCE in $DB_INSTANCES; do
-      aws rds delete-db-instance --db-instance-identifier "${INSTANCE}" --skip-final-snapshot > /dev/null 2>&1
-    done
+    if [ -z "${DRY_RUN}" ]; then
+      echo "Deleting db instances for $1: $DB_INSTANCES"
+      for INSTANCE in $DB_INSTANCES; do
+        echo "Deleting db instance: $INSTANCE"
+        aws rds delete-db-instance --db-instance-identifier "${INSTANCE}" --skip-final-snapshot > /dev/null 2>&1
+      done
+    else
+      echo "DB Instances that will be deleted: $DB_INSTANCES"
+      echo "DB Instance Count:$(echo $DB_INSTANCES |xargs -n1 echo | wc -l)"
+    fi
   fi
 
   #Search for DB clusters and delete them
@@ -35,11 +57,17 @@ delete_db_resources () {
   if [ "${CLUSTERS}" = "" ];then
     echo "No db clusters found with prefix $1. Nothing to delete."
   else
-    echo "Deleting db clusters for $1:"
-    for CLUSTER in $CLUSTERS; do
-      aws rds delete-db-cluster --db-cluster-identifier "$CLUSTER" --skip-final-snapshot > /dev/null 2>&1
-      aws rds wait db-cluster-deleted --db-cluster-identifier "$CLUSTER"
-    done
+    if [ -z "${DRY_RUN}" ]; then
+      echo "Deleting db clusters for $1: ${CLUSTERS}"
+      for CLUSTER in $CLUSTERS; do
+        echo "Deleting cluster: $CLUSTER"
+        aws rds delete-db-cluster --db-cluster-identifier "$CLUSTER" --skip-final-snapshot > /dev/null 2>&1
+        aws rds wait db-cluster-deleted --db-cluster-identifier "$CLUSTER"
+      done
+    else
+      echo "DB Clusters that will be deleted: $CLUSTERS"
+      echo "DB Clusters Count:$(echo $CLUSTERS |xargs -n1 echo | wc -l)"
+    fi
   fi
 
   #Search for DB snapshots and delete them
@@ -49,10 +77,16 @@ delete_db_resources () {
   if [ "${SNAPSHOTS}" = "" ];then
     echo "No db snapshots found with prefix $1. Nothing to delete."
   else
-    echo "Deleting db snapshots for $1:"
-    for SNAPSHOT in $SNAPSHOTS; do
-      aws rds delete-db-snapshot --db-snapshot-identifier "$SNAPSHOT" > /dev/null 2>&1
-    done
+    if [ -z "${DRY_RUN}" ]; then
+      echo "Deleting db snapshots for $1: ${SNAPSHOTS}"
+      for SNAPSHOT in $SNAPSHOTS; do
+        echo "Deleting db snapshot: $SNAPSHOT"
+        aws rds delete-db-snapshot --db-snapshot-identifier "$SNAPSHOT" > /dev/null 2>&1
+      done
+    else
+      echo "DB Snapshots that will be deleted: $SNAPSHOTS"
+      echo "DB Snapshots Count:$(echo $SNAPSHOTS |xargs -n1 echo | wc -l)"
+    fi
   fi
 }
 
@@ -65,12 +99,17 @@ delete_lb_resources () {
   if [ "${LB_ARN_LIST}" = "" ];then
     echo "No load balancers found with prefix $1. Nothing to delete."
   else
-    echo "Deleting load balancers for $1:"
-    #Loop through the load balancer ARNs and delete the load balancers
-    for LB_ARN in $LB_ARN_LIST; do
-      echo "Deleting load balancer $LB_ARN"
-      aws elbv2 delete-load-balancer --load-balancer-arn "$LB_ARN"
-    done
+    if [ -z "${DRY_RUN}" ]; then
+      echo "Deleting load balancers for $1: ${LB_ARN_LIST}"
+      #Loop through the load balancer ARNs and delete the load balancers
+      for LB_ARN in $LB_ARN_LIST; do
+        echo "Deleting load balancer $LB_ARN"
+        aws elbv2 delete-load-balancer --load-balancer-arn "$LB_ARN"
+      done
+    else
+      echo "Load Balancers that will be deleted: $LB_ARN_LIST"
+      echo "Load Balancers Count:$(echo $LB_ARN_LIST |xargs -n1 echo | wc -l)"
+    fi
   fi
 }
 
@@ -83,12 +122,17 @@ delete_target_groups () {
   if [ "${TG_ARN_LIST}" = "" ];then
     echo "No target groups found with prefix $1. Nothing to delete."
   else
-    echo "Deleting target groups for $1:"
-    #Loop through the target group ARNs and delete the target groups
-    for TG_ARN in $TG_ARN_LIST; do
-      echo "Deleting target group $TG_ARN"
-      aws elbv2 delete-target-group --target-group-arn "$TG_ARN"
-    done
+    if [ -z "${DRY_RUN}" ]; then
+      echo "Deleting target groups for $1: ${TG_ARN_LIST}"
+      #Loop through the target group ARNs and delete the target groups
+      for TG_ARN in $TG_ARN_LIST; do
+        echo "Deleting target group $TG_ARN"
+        aws elbv2 delete-target-group --target-group-arn "$TG_ARN"
+      done
+    else
+      echo "Target Groups that will be deleted: $TG_ARN_LIST"
+      echo "Target Groups Count:$(echo $TG_ARN_LIST |xargs -n1 echo | wc -l)"
+    fi
   fi
 }
 
@@ -106,36 +150,49 @@ delete_route53 () {
   #Get ResourceRecord Value
   RECORD_VALUE=$(aws route53 list-resource-record-sets \
     --hosted-zone-id "${R53_ZONE_ID}" \
-    --query "ResourceRecordSets[?starts_with(Name, '${NAME_PREFIX_LOWER}.') \
-    && Type == 'CNAME'].ResourceRecords[0].Value" --output text)
-
-
-  #Delete Route53 record
-  if [[ "$R53_RECORD" == "${NAME_PREFIX_LOWER}."* ]]; then
-    echo "Deleting Route53 record ${R53_RECORD} for prefix $1"
-    CHANGE_STATUS=$(aws route53 change-resource-record-sets --hosted-zone-id "${R53_ZONE_ID}" \
-    --change-batch '{"Changes": [
-            {
-                "Action": "DELETE",
-                "ResourceRecordSet": {
-                    "Name": "'"${R53_RECORD}"'",
-                    "Type": "CNAME",
-                    "TTL": 300,
-                    "ResourceRecords": [
-                        {
-                            "Value": "'"${RECORD_VALUE}"'"
-                        }
-                    ]
-                }
-            }
-        ]
-    }')
-    STATUS_ID=$(echo "$CHANGE_STATUS" | jq -r '.ChangeInfo.Id')
-    #Get status from the change
-    aws route53 wait resource-record-sets-changed --id "$STATUS_ID"
-    echo "Successfully deleted Route53 record ${R53_RECORD}: status: ${STATUS_ID}"
+    --query "ResourceRecordSets[?starts_with(Name, '${NAME_PREFIX_LOWER}')].ResourceRecords[0].Value" --output text)
+  echo "Route53 RECORD_VALUE:
+  $RECORD_VALUE"
+  RECORD_COUNT=$(echo $RECORD_VALUE | xargs -n1 echo | wc -l)
+  echo "Total Route53 Record COUNT is: $RECORD_COUNT"
+  if [ "${RECORD_COUNT}" == 0 ]; then
+    echo "No Route53 records found for prefix $1. Nothing to delete."
   else
-    echo "No Route53 record found for prefix $1. Nothing to delete."
+    if [ -z "${DRY_RUN}" ]; then
+      for i in $(echo $RECORD_VALUE | xargs -n1 echo)
+      do 
+          NAME="$(echo "$i" | cut -d "-" -f1)-r53.qa.rancher.space."
+          VALUE=$i
+          echo "
+{\"Changes\": [
+        {
+            \"Action\": \"DELETE\",
+            \"ResourceRecordSet\": {
+                \"Name\": \"${NAME}\",
+                \"Type\": \"CNAME\",
+                \"TTL\": 300,
+                \"ResourceRecords\": [
+                    {
+                        \"Value\": \"${VALUE}\"
+                    }
+                ]
+            }
+        }
+    ]
+}" > payload.json
+        cat payload.json
+        echo "Deleting Route53 record ${NAME} for prefix $1"
+        CHANGE_STATUS=$(aws route53 change-resource-record-sets --hosted-zone-id "${R53_ZONE_ID}" --change-batch file://"${PWD}"/payload.json)
+        echo "CHANGE STATUS: 
+        $CHANGE_STATUS"
+        STATUS_ID=$(echo "$CHANGE_STATUS" | grep Id: | cut -d " " -f 4)
+        echo "Status ID: $STATUS_ID"
+        #Get status from the change
+        aws route53 wait resource-record-sets-changed --id "$STATUS_ID"
+        echo "Successfully deleted Route53 record ${NAME}: status: ${STATUS_ID}"
+      done
+      rm -rf "${PWD}"/payload.json
+    fi
   fi
 }
 
@@ -153,23 +210,28 @@ do
   case "${OPTION}"
     in
     r) RESOURCES=${OPTARG};;
+    d) DRY_RUN="true";;
     h|?)
       echo "
     Usage: 
-      $(basename "$0") [-r <resource_prefix>] [-h]
+      $(basename "$0") [-r <resource_prefix>] [-d] [-h]
 
       -r: resource prefix names. can be comma separated. 
+      -d: Dry run option. This will simply list the names and ids of what could potentially be listed. But will not delete anything.
       -h: help. Prints usage example.
       
       Ex 1: 
-      ./delete_jenkins_resources.sh -r resourceprefix1,resourceprefix2,resourceprefix2
+      ./delete_resources.sh -r resourceprefix1,resourceprefix2,resourceprefix2
       - This will delete all resources for the names resourceprefix1,resourceprefix2 and resourceprefix2
       Ex 2:
-      ./delete_jenkins_resources.sh
+      ./delete_resources.sh
       - This will lookup the resource name from local tfvars file and delete the resources. This is interactive and will ask for confirmation before delete
       Ex 3: 
-      ./delete_jenkins_resources.sh -h
+      ./delete_resources.sh -h
       - Print usage details for reference.
+      Ex 4:
+      ./delete_resources.sh -r resourceprefix1 -d
+      - This is a dry run. You will see results of what may potentially get deleted but will not do any actual deletion.
       "
       exit 1
       ;;
@@ -189,13 +251,19 @@ if [ "${RESOURCES}" = "" ]; then
       # Split string based on delimiter :
       PRODUCT_NAME=$(echo "${PRODUCT_NAME}" | cut -d ":" -f 2)  
     fi
+    echo "PRODUCT NAME is: $PRODUCT_NAME"
+  
+    if echo "$PRODUCT_NAME" | grep -q "k3s" && echo "$PRODUCT_NAME" | grep -q "rke2"; then
+      echo "Please have ONLY ONE product name. Either k3s or rke2. Exiting."
+      exit 1
+    fi
     if [[ -z "$PRODUCT_NAME" || ! "$PRODUCT_NAME" =~ ^(rke2|k3s)$ ]]; then
       echo "Wrong or empty product name found in .env file for: $PRODUCT_NAME"
       exit 1
     fi
 
     #Validate path to the tfvars file
-    if [[ ! -f ./config/"$PRODUCT_NAME".tfvars ]]; then
+    if [[ ! -f "${$CONFIG_DIR}"/"$PRODUCT_NAME".tfvars ]]; then
       echo "No $PRODUCT_NAME.tfvars file found in config directory"
       exit 1
     fi
@@ -218,7 +286,13 @@ if [ "${RESOURCES}" = "" ]; then
 else
     for i in $(echo "${RESOURCES}" | tr "," "\n")
     do
-      echo "## For prefix name: $i:"
-      delete_all_resources "$i"
+      PREFIX_LENGTH=${#i}
+      if [ "$PREFIX_LENGTH" -gt 5 ]; then
+        echo "## For prefix name: $i:"
+        delete_all_resources "$i"
+      else
+        echo "Length of prefix name $i lesser than 5. Please provide a tangible prefix length for deletion."
+        exit 1
+      fi
     done
 fi
